refactor : 채팅 컴포넌트, 페이지 정리
