+++ src/components/chat/ChatList.tsx
@@ -4,7 +4,7 @@ import ProfileIMG from '../assets/images/profile.png'
 import ChatIMG from '../assets/images/chatIcon.png'
 import { FaCircle } from 'react-icons/fa6'
 import { RxDividerVertical } from 'react-icons/rx'
-import { useApiUrlStore, useChatListStore } from '../store/store'
+import { useApiUrlStore, useChatListStore } from '../../store/store'
 import axios from 'axios'
 import { useEffect } from 'react'
 import { Link } from 'react-router-dom'
